rename project to gfx-php
