+++ README.md
@@ -1,8 +1,8 @@
-# image-php - The pure PHP image processing library
+# gfx-php - The pure PHP graphics library
 
-[![Build Status](https://travis-ci.org/mike42/image-php.svg?branch=master)](https://travis-ci.org/mike42/image-php) [![Latest Stable Version](https://poser.pugx.org/mike42/image-php/v/stable)](https://packagist.org/packages/mike42/image-php)
-[![Total Downloads](https://poser.pugx.org/mike42/image-php/downloads)](https://packagist.org/packages/mike42/image-php)
-[![License](https://poser.pugx.org/mike42/image-php/license)](https://packagist.org/packages/mike42/image-php)
+[![Build Status](https://travis-ci.org/mike42/gfx-php.svg?branch=master)](https://travis-ci.org/mike42/gfx-php) [![Latest Stable Version](https://poser.pugx.org/mike42/gfx-php/v/stable)](https://packagist.org/packages/mike42/gfx-php)
+[![Total Downloads](https://poser.pugx.org/mike42/gfx-php/downloads)](https://packagist.org/packages/mike42/gfx-php)
+[![License](https://poser.pugx.org/mike42/gfx-php/license)](https://packagist.org/packages/mike42/gfx-php)
 
 This library implements input, output and processing of raster images in pure PHP, so that image
 processing extensions (Gd, Imagick) are not required.
