+++ src/Mike42/GfxPhp/Codec/BmpCodec.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Mike42\ImagePhp\Codec;
+namespace Mike42\GfxPhp\Codec;
 
-use Mike42\ImagePhp\RasterImage;
-use Mike42\ImagePhp\RgbRasterImage;
+use Mike42\GfxPhp\RasterImage;
+use Mike42\GfxPhp\RgbRasterImage;
 
 class BmpCodec implements ImageEncoder
 {
