+++ src/Mike42/GfxPhp/Codec/GifCodec.php
@@ -1,9 +1,9 @@
 <?php
-namespace Mike42\ImagePhp\Codec;
+namespace Mike42\GfxPhp\Codec;
 
-use Mike42\ImagePhp\RasterImage;
-use Mike42\ImagePhp\Util\LzwCompression;
-use Mike42\ImagePhp\IndexedRasterImage;
+use Mike42\GfxPhp\RasterImage;
+use Mike42\GfxPhp\Util\LzwCompression;
+use Mike42\GfxPhp\IndexedRasterImage;
 
 class GifCodec implements ImageEncoder
 {
