+++ src/Mike42/GfxPhp/Codec/ImageDecoder.php
@@ -1,7 +1,7 @@
 <?php
-namespace Mike42\ImagePhp\Codec;
+namespace Mike42\GfxPhp\Codec;
 
-use Mike42\ImagePhp\RasterImage;
+use Mike42\GfxPhp\RasterImage;
 
 interface ImageDecoder
 {
