+++ src/Mike42/GfxPhp/IndexedRasterImage.php
@@ -1,5 +1,5 @@
 <?php
-namespace Mike42\ImagePhp;
+namespace Mike42\GfxPhp;
 
 class IndexedRasterImage extends AbstractRasterImage
 {
